// src/app/core/api/generated/schemas/mediaGalleryItemDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { MediaGalleryItemDtoAltText } from "./mediaGalleryItemDtoAltText";
import type { MediaGalleryItemDtoCreatedBy } from "./mediaGalleryItemDtoCreatedBy";

export interface MediaGalleryItemDto {
  media_asset_id: number;
  media_type: string;
  media_url: string;
  /** @nullable */
  alt_text?: MediaGalleryItemDtoAltText;
  is_active: boolean;
  created_at: string;
  created_by?: MediaGalleryItemDtoCreatedBy;
}